fullstack-orthanc-dicom/auth-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
flask-limiter==3.5.0
